+++ src/Repository/UrlHistoryRepository.php
@@ -5,15 +5,22 @@ declare(strict_types=1);
 namespace Umanit\SeoBundle\Repository;
 
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Umanit\SeoBundle\UrlHistory\UrlPoolerInterface;
+use Umanit\SeoBundle\Entity\UrlHistory;
 
-class UrlHistoryRepository extends ServiceEntityRepository
+class UrlHistoryRepository extends ServiceEntityRepository implements UrlHistoryRepositoryInterface
 {
-    /**
-     * @return array<int, UrlPoolerInterface>
-     */
     public function findBySeoUuid(string $seoUuid): array
     {
         return $this->findBy(['seoUuid' => $seoUuid], ['id' => 'ASC']);
+    }
+
+    public function findOneByOldPath(string $oldPath, string $locale): ?UrlHistory
+    {
+        return $this->findOneBy(['oldPath' => $oldPath, 'locale' => $locale]);
+    }
+
+    public function findOneByOldPathAndSeoUuid(string $oldPath, string $seoUuid, string $locale): ?UrlHistory
+    {
+        return $this->findOneBy(['oldPath' => $oldPath, 'seoUuid' => $seoUuid, 'locale' => $locale]);
     }
 }